+++ src/gui/wizard/owncloudsetuppage.cpp
@@ -116,11 +116,11 @@ void OwncloudSetupPage::slotUrlChanged(const QString& url)
     }
 
     if (url.startsWith(QLatin1String("http://"))) {
-        _ui.urlLabel->setPixmap(QPixmap(Utility::hidpiFileName(":/client/resources/lock-http.png")));
+        _ui.urlLabel->setPixmap(QPixmap(Theme::hidpiFileName(":/client/resources/lock-http.png")));
         _ui.urlLabel->setToolTip(tr("This url is NOT secure as it is not encrypted.\n"
                                     "It is not advisable to use it."));
     } else {
-        _ui.urlLabel->setPixmap(QPixmap(Utility::hidpiFileName(":/client/resources/lock-https.png")));
+        _ui.urlLabel->setPixmap(QPixmap(Theme::hidpiFileName(":/client/resources/lock-https.png")));
         _ui.urlLabel->setToolTip(tr("This url is secure. You can use it."));
     }
 }
